Large-scale OS update
Increase app reliability, add connful, introduce blank user privileges, add uptime command, allow for automatic configuration

+++ os_system/combine_into_index/apps/power.js
@@ -5,10 +5,16 @@
 (async function() {
 	// @pcos-app-mode isolatable
 	await availableAPIs.windowVisibility(false);
+	if (!(await availableAPIs.getPrivileges()).includes("GET_LOCALE")) { await availableAPIs.toMyCLI("power: Locale permission denied\r\n");
+		return await availableAPIs.terminate();	}
 	await availableAPIs.attachCLI();
-	await availableAPIs.shutdown({
-		isReboot: exec_args.includes("r") || exec_args.includes("reboot") || exec_args.includes("restart"),
-	});
+	try {
+		await availableAPIs.shutdown({
+			isReboot: exec_args.includes("r") || exec_args.includes("reboot") || exec_args.includes("restart"),
+		});
+	} catch (e) {
+		await availableAPIs.toMyCLI("power: " + await availableAPIs.lookupLocale(e.message) + "\r\n");
+	}
 	await availableAPIs.terminate();
 })();
